requirements.txt
tensorflow
opencv-python
numpy
pillow
matplotlib
streamlit
scikit-learn